4.15 - 4.21 Blogs

+++ part4/blog/controllers/blogs.js
@@ -2,9 +2,13 @@ const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
 const User = require("../models/user");
 const { error } = require("../utils/logger");
+const jwt = require("jsonwebtoken");
 
 blogsRouter.get("/", async (req, res) => {
-    const blogs = await Blog.find({}).populate('user', {username: 1, name: 1});
+    const blogs = await Blog.find({}).populate("user", {
+        username: 1,
+        name: 1,
+    });
     res.json(blogs);
 });
 
@@ -16,8 +20,14 @@ blogsRouter.get("/:id", async (req, res, next) => {
 
 blogsRouter.post("/", async (req, res, next) => {
     const body = req.body;
+    const decodedToken = jwt.verify(req.token, process.env.SECRET);
+    if (!decodedToken.id) {
+        return res.status(401).json({
+            error: "token invalid",
+        });
+    }
+    const user = await User.findById(decodedToken.id);
 
-    const user = await User.findById(body.userID);
     if (!user) {
         return res.status(400).json({ error: "userID missing or not valid" });
     }
@@ -26,20 +36,37 @@ blogsRouter.post("/", async (req, res, next) => {
         author: body.author,
         url: body.url,
         likes: body.likes,
-        user: user._id
+        user: user._id,
     });
 
     const savedBlog = await blog.save();
-    user.blogs = user.blogs.concat(savedBlog._id)
-    await user.save()
-    
+    user.blogs = user.blogs.concat(savedBlog._id);
+    await user.save();
+
     res.status(201).json(savedBlog);
 });
-
 blogsRouter.delete("/:id", async (req, res, next) => {
+    const decodedToken = jwt.verify(req.token, process.env.SECRET);
+    if (!decodedToken.id) {
+        return res.status(401).json({
+            error: "token invalid",
+        });
+    }
+    const user = await User.findById(decodedToken.id);
+    if (!user) {
+        return res.status(400).json({ error: "userID missing or not valid" });
+    }
+    const blog = await Blog.findById(req.params.id);
+    if (!blog || decodedToken.id !== blog.user.toString()) {
+        return res.status(404).json({ error: "blog id dont exist" });
+    }
     await Blog.findByIdAndDelete(req.params.id);
     res.status(204).end();
 });
+// blogsRouter.delete("/:id", async (req, res, next) => {
+//     await Blog.findByIdAndDelete(req.params.id);
+//     res.status(204).end();
+// });
 
 blogsRouter.put("/:id", async (req, res, next) => {
     const { likes } = req.body;
